Crate sp_mmr_primitives

Merkle Mountain Range primitive types.

Re-exports

pub use mmr_lib;

Modules

utils

Merkle Mountain Range utilities.

Structs

AncestryProof

An MMR ancestry proof for a prior mmr root.

Compact

A composition of multiple leaf elements with compact form representation.

EncodableOpaqueLeaf

A type-safe wrapper for the concrete leaf type.

LeafProof

An MMR proof data for a group of leaves.

OpaqueLeaf

A helper type to allow using arbitrary SCALE-encoded leaf data in the RuntimeApi.

Enums

DataOrHash

An element representing either full data or its hash.

Error

Merkle Mountain Range operation error.

Constants

INDEXING_PREFIX

Prefix for elements stored in the Off-chain DB via Indexing API.

Traits

FullLeaf

A full leaf content stored in the offchain-db.

LeafDataProvider

A provider of the MMR’s leaf data.

MmrApi

API to interact with MMR pallet.

OnNewRoot

New MMR root notification hook.

Type Aliases

LeafIndex

A type to describe leaf position in the MMR.

NodeIndex

A type to describe node position in the MMR (node index).