Crate sp_mmr_primitives

Merkle Mountain Range primitive types.

Structs

BatchProof

A MMR proof data for a group of leaves.

Compact

A composition of multiple leaf elements with compact form representation.

EncodableOpaqueLeaf

A type-safe wrapper for the concrete leaf type.

OpaqueLeaf

A helper type to allow using arbitrary SCALE-encoded leaf data in the RuntimeApi.

Proof

A MMR proof data for one of the leaves.

Enums

DataOrHash

An element representing either full data or its hash.

Error

Merkle Mountain Range operation error.

Traits

FullLeaf

A full leaf content stored in the offchain-db.

LeafDataProvider

A provider of the MMR’s leaf data.

MmrApi

API to interact with MMR pallet.

OnNewRoot

New MMR root notification hook.

Type Definitions

LeafIndex

A type to describe leaf position in the MMR.

NodeIndex

A type to describe node position in the MMR (node index).